requirements.txt
certifi==2021.5.30
cycler==0.10.0
Cython==0.29.24
cytoolz==0.11.0
decorator==5.1.0
GPy==1.10.0
GPyOpt==1.2.6
imageio==2.9.0
joblib==1.0.1
kiwisolver==1.3.2
lime==0.2.0.1
matplotlib==3.3.2
networkx==2.6.3
numpy==1.21.2
pandas==1.1.3
paramz==0.9.5
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.5.2
six==1.16.0
threadpoolctl==2.2.0
tifffile==2021.8.30
toolz==0.11.1
tqdm==4.50.2
wincertstore==0.2
